app/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
google-ai-generativelanguage==0.6.15 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.168.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.39.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.31.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
google-generativeai==0.8.5 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.62.3 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.7 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.3 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
